# project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/employesNotRelated": {
            "get": {
                "tags": [
                    "Employes"
                ],
                "summary": "List Employes whos are not related",
                "responses": {
                    "200": {
                        "description": "Success: List all Employes who are not related",
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/employes": {
            "get": {
                "tags": [
                    "Employes"
                ],
                "summary": "List Employes",
                "responses": {
                    "200": {
                        "description": "Success: List all Employes",
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employes"
                ],
                "summary": "Create Employe",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Employe",
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/employes/{id}": {
            "get": {
                "tags": [
                    "Employes"
                ],
                "summary": "Get Employe by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Employe by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Employe",
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employes"
                ],
                "summary": "Update Employe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Employe by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Employe updated",
                        "schema": {
                            "$ref": "#/definitions/Employe"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employes"
                ],
                "summary": "Delete Employe",
                "description": "Delete the specified Employe by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employe id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Employe": {
            "required": [
                "nom",
                "prenom",
                "dateNaiss",
                "adresse_id"
            ],
            "properties": {
                "nom": {
                    "description": "nom de l'employ\u00e9",
                    "type": "string",
                    "example": "VLAMISKOT, DJEALE"
                },
                "prenom": {
                    "description": "prenom de l'employ\u00e9",
                    "type": "string",
                    "example": "komi, Jack, Jessica"
                },
                "dateNaiss": {
                    "description": "date de naissance de l'employ\u00e9",
                    "type": "Date",
                    "example": "2009/23/05, 2008/23/11, 2007/09/06"
                },
                "adresse_id": {
                    "description": "Association with: adresse",
                    "type": "integer",
                    "example": "2"
                }
            }
        },
...
    }
}